Removed "player" variable as this lead to several issues
Created a player variable to store the current player however this lead to several minor issues as the player is constantly being updated and the variable was not updated simultaneously. This is why I removed it and used game.getPlayer() instead to ensure the most current version is always used.